refactor(formats): move the `TableProxy` object to formats from the operations
